requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
black==24.10.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
Deprecated==1.2.15
durationpy==0.9
fastapi==0.109.2
filelock==3.16.1
flake8==7.1.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
ghp-import==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.26.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
langchain==0.1.9
langchain-core==0.1.27
langchain-text-splitters==0.3.3
langsmith==0.2.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.49
mkdocs-material-extensions==1.3.1
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy==1.13.0
mypy-extensions==1.0.0
narwhals==1.18.4
numpy==2.2.0
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
paginate==0.5.7
pandas==2.2.1
pathspec==0.12.1
pgvector==0.3.6
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.7.4
propcache==0.2.1
protobuf==5.29.2
psycopg2-binary==2.9.9
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pymdown-extensions==10.12
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.11.1
python-socketio-client==0.9.5
python-socketio-server==0.9.5
pytz==2024.2
PyYAML==6.0.2
pyyaml_env_tag==0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.27.1
watchdog==6.0.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==12.0
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
networkx==3.2.1
neo4j==5.18.0
elevenlabs==0.2.24
aiofiles==23.2.1
